requirements.txt
Flask==2.2.2
fpdf==1.7.2
mysql_connector_repackaged==0.3.1
opencv_python==4.6.0.66
Pillow==9.2.0
PyPDF4==1.27.0
gunicorn